requirements.txt
streamlit
python-docx
groq      
python-dotenv
pdfplumber
scikit-learn
